requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.9
deep-translator==1.10.1
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.1
httpx==0.24.1
huggingface-hub==0.14.1
hyperframe==6.0.1
charset-normalizer==3.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
langchain==0.0.234
langsmith==0.0.5
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pip==23.1.2
portalocker==2.7.0
protobuf==4.23.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
qdrant-client==1.1.7
regex==2023.5.5
requests==2.30.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
termcolor==2.3.0
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing_extensions==4.5.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.15
uuid==1.30
Werkzeug==2.3.4
yarl==1.9.2
